Ajout d'une condition dans le script pour un rm

+++ stats-for-race.sh
@@ -151,4 +151,6 @@ else
 fi
 echo ""
 
-rm stats.log
+if [ -e stats.log ]; then
+    rm stats.log
+fi
